Grounded-AI/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
faiss-cpu
sentence-transformers
pypdf
groq
openai
pandas
numpy
